feat(myapp.cpp): multithreading
compiles + runs, doesn't do colors right
